fix: separate several files from handler_analyzer.go
