+++ internal/extractor/gin/handler_analyzer.go
@@ -36,68 +36,6 @@ func NewHandlerAnalyzer(fset *token.FileSet, files map[string]*ast.File) *Handle
 	return a
 }
 
-// discoverHelperFuncs scans all parsed files for package-level functions
-// that accept *gin.Context as their first parameter. These are potential
-// response helpers that can be traced for cross-function response extraction.
-// Helpers are stored per-package to avoid cross-package name collisions.
-func (a *HandlerAnalyzer) discoverHelperFuncs() {
-	for _, file := range a.files {
-		pkg := file.Name.Name
-		for _, decl := range file.Decls {
-			fn, ok := decl.(*ast.FuncDecl)
-			if !ok || fn.Recv != nil || fn.Type.Params == nil || len(fn.Type.Params.List) == 0 {
-				continue
-			}
-			// Check if first parameter is *gin.Context
-			firstParam := fn.Type.Params.List[0]
-			if isGinContextType(firstParam.Type) {
-				if a.helperFuncs[pkg] == nil {
-					a.helperFuncs[pkg] = make(map[string]*ast.FuncDecl)
-				}
-				a.helperFuncs[pkg][fn.Name.Name] = fn
-			}
-		}
-	}
-	count := 0
-	for _, funcs := range a.helperFuncs {
-		count += len(funcs)
-	}
-	if count > 0 {
-		slog.Debug("Discovered potential helper functions", "count", count)
-	}
-}
-
-// pkgForFunc determines which package a function declaration belongs to
-// by searching all parsed files for a pointer match.
-func (a *HandlerAnalyzer) pkgForFunc(fn *ast.FuncDecl) string {
-	for _, file := range a.files {
-		for _, decl := range file.Decls {
-			if f, ok := decl.(*ast.FuncDecl); ok && f == fn {
-				return file.Name.Name
-			}
-		}
-	}
-	return ""
-}
-
-// lookupHelper finds a helper function by name, preferring the caller's package
-// scope and falling back to a global search across all packages.
-func (a *HandlerAnalyzer) lookupHelper(callerPkg, name string) *ast.FuncDecl {
-	// Try caller's package first
-	if pkgFuncs, ok := a.helperFuncs[callerPkg]; ok {
-		if fn, ok := pkgFuncs[name]; ok {
-			return fn
-		}
-	}
-	// Fallback: search all packages
-	for _, pkgFuncs := range a.helperFuncs {
-		if fn, ok := pkgFuncs[name]; ok {
-			return fn
-		}
-	}
-	return nil
-}
-
 // isGinContextType checks if an AST expression represents *gin.Context.
 func isGinContextType(expr ast.Expr) bool {
 	star, ok := expr.(*ast.StarExpr)
@@ -607,352 +545,5 @@ func (a *HandlerAnalyzer) extractFileParam(call *ast.CallExpr, info *HandlerInfo
 			GoType:   "file",
 			Required: true,
 		})
-	}
-}
-
-// extractResponse extracts response information from c.JSON(), c.XML(), etc. calls.
-func (a *HandlerAnalyzer) extractResponse(call *ast.CallExpr, info *HandlerInfo, varTypeMap map[string]string, goType string) {
-	if len(call.Args) < 2 {
-		return
-	}
-	statusCode := extractStatusCode(call.Args[0])
-	if goType == "" {
-		// Pass status code to determine if we should unwrap Data field
-		goType = extractTypeFromResponseWithStatus(call.Args[1], varTypeMap, statusCode)
-	}
-	if goType != "" {
-		slog.Debug("Extracted response type", "status", statusCode, "type", goType)
-	}
-	info.Responses = append(info.Responses, ResponseInfo{
-		StatusCode: statusCode,
-		GoType:     goType,
-	})
-}
-
-// extractHelperResponse analyzes helper call arguments to extract response types.
-// Patterns:
-//   - done(c, err)           → only default error response
-//   - done(c, data)          → 200 success with data type
-//   - done(c, data, err)     → 200 success + default error
-func (a *HandlerAnalyzer) extractHelperResponse(args []ast.Expr, info *HandlerInfo, varTypeMap map[string]string) {
-	// Skip first arg (*gin.Context)
-	dataArgs := args[1:]
-
-	for _, arg := range dataArgs {
-		if ident, ok := arg.(*ast.Ident); ok && ident.Name == "nil" {
-			continue
-		}
-
-		goType := extractTypeFromResponseInternal(arg, varTypeMap, 200, true)
-		if goType == "" {
-			continue
-		}
-
-		// Check if this looks like an error type (err variable, *gin.Error, etc.)
-		if isErrorType(arg, varTypeMap) {
-			// Generate default error response
-			info.Responses = append(info.Responses, ResponseInfo{
-				StatusCode: 0, // 0 = default
-				GoType:     goType,
-			})
-			slog.Debug("Extracted helper error response", "type", goType)
-		} else {
-			// Generate 200 success response
-			info.Responses = append(info.Responses, ResponseInfo{
-				StatusCode: 200,
-				GoType:     goType,
-			})
-			slog.Debug("Extracted helper success response", "type", goType)
-		}
-	}
-}
-
-// parseHelperCall detects standalone helper function calls like done(c, data, err)
-// and extracts response types from the arguments.
-// Unlike parseHandlerCall which handles c.XXX() method calls, this handles
-// bare function calls where the first arg is *gin.Context.
-//
-// Two detection strategies:
-//  1. Known helpers (done, respond, etc.) — extract types from call arguments directly
-//  2. Discovered helpers — trace into the function body to find c.JSON/c.XML calls
-func (a *HandlerAnalyzer) parseHelperCall(call *ast.CallExpr, info *HandlerInfo, varTypeMap map[string]string) {
-	// Only handle simple function calls: done(...), not c.XXX() or pkg.Fn()
-	ident, ok := call.Fun.(*ast.Ident)
-	if !ok {
-		return
-	}
-
-	// Need at least 2 args: c + something
-	if len(call.Args) < 2 {
-		return
-	}
-
-	if isKnownResponseHelper(ident.Name) {
-		a.extractHelperResponse(call.Args, info, varTypeMap)
-		return
-	}
-
-	// Cross-function call tracking: trace into discovered helper functions
-	if helper := a.lookupHelper(a.callerPkg, ident.Name); helper != nil {
-		a.traceHelperResponse(helper, call.Args, info, varTypeMap)
-	}
-}
-
-// traceHelperResponse traces into a helper function's body to find c.JSON/c.XML
-// response calls and extracts response types by mapping call-site arguments to
-// the helper's parameters.
-func (a *HandlerAnalyzer) traceHelperResponse(helperDecl *ast.FuncDecl, callArgs []ast.Expr, info *HandlerInfo, varTypeMap map[string]string) {
-	if helperDecl.Body == nil {
-		return
-	}
-
-	// Build param-to-argument mapping for the helper function
-	// e.g., helper(c *gin.Context, data any, err error) called as helper(c, project, nil)
-	// maps: data → project (with type from varTypeMap), err → nil
-	paramArgMap := a.buildParamArgMap(helperDecl, callArgs, varTypeMap)
-
-	// Scan helper body for c.JSON/c.XML calls
-	ast.Inspect(helperDecl.Body, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			return true
-		}
-		sel, ok := call.Fun.(*ast.SelectorExpr)
-		if !ok {
-			return true
-		}
-		if _, ok = sel.X.(*ast.Ident); !ok {
-			return true
-		}
-
-		method := sel.Sel.Name
-		switch method {
-		case "JSON", "XML", "YAML":
-			if len(call.Args) >= 2 {
-				// Resolve the response argument through param mapping
-				resolvedArg := a.resolveThroughParamMap(call.Args[1], paramArgMap)
-				statusCode := extractStatusCode(call.Args[0])
-				goType := extractTypeFromResponseWithStatus(resolvedArg, varTypeMap, statusCode)
-				if goType != "" {
-					info.Responses = append(info.Responses, ResponseInfo{
-						StatusCode: statusCode,
-						GoType:     goType,
-					})
-					slog.Debug("Traced helper response", "helper", helperDecl.Name.Name, "status", statusCode, "type", goType)
-				}
-			}
-		}
-		return true
-	})
-}
-
-// buildParamArgMap maps helper function parameter names to their call-site arguments.
-// e.g., for helper(c *gin.Context, data any, err error) called as helper(c, project, nil):
-//
-//	"data" → *ast.Ident{Name: "project"}, "err" → *ast.Ident{Name: "nil"}
-func (a *HandlerAnalyzer) buildParamArgMap(helperDecl *ast.FuncDecl, callArgs []ast.Expr, _ map[string]string) map[string]ast.Expr {
-	paramArgMap := make(map[string]ast.Expr)
-	if helperDecl.Type.Params == nil {
-		return paramArgMap
-	}
-
-	params := helperDecl.Type.Params.List
-	for i, param := range params {
-		if i >= len(callArgs) {
-			break
-		}
-		// Skip the first param (*gin.Context) — it's always the context
-		if i == 0 {
-			continue
-		}
-		for _, name := range param.Names {
-			paramArgMap[name.Name] = callArgs[i]
-		}
-	}
-	return paramArgMap
-}
-
-// resolveThroughParamMap resolves an expression by substituting helper parameters
-// with their call-site arguments. If the expression is an Ident matching a helper
-// parameter name, the corresponding call-site argument is returned instead.
-func (a *HandlerAnalyzer) resolveThroughParamMap(expr ast.Expr, paramToArg map[string]ast.Expr) ast.Expr {
-	switch e := expr.(type) {
-	case *ast.Ident:
-		if arg, exists := paramToArg[e.Name]; exists {
-			return arg
-		}
-	case *ast.CompositeLit:
-		for i, elt := range e.Elts {
-			e.Elts[i] = a.resolveThroughParamMap(elt, paramToArg)
-		}
-	case *ast.KeyValueExpr:
-		e.Value = a.resolveThroughParamMap(e.Value, paramToArg)
-	}
-	return expr
-}
-
-// inferParamTypes scans the handler body for strconv conversions and conditional
-// comparisons that reveal the true type of query/path/header parameters previously
-// extracted as string.
-//
-// Detected patterns:
-//   - strconv.Atoi(c.Query("offset"))  → integer
-//   - strconv.ParseInt(c.Query("id"), ...)  → integer
-//   - strconv.ParseBool(c.Query("verbose"))  → boolean
-//   - strconv.ParseFloat(c.Query("rate"), ...)  → number
-//   - c.Query("x") == "true"  → boolean
-func (a *HandlerAnalyzer) inferParamTypes(body *ast.BlockStmt, info *HandlerInfo) {
-	ast.Inspect(body, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			// Also check for conditional bool inference: c.Query("x") == "true"
-			a.inferBoolFromComparison(n, info)
-			return true
-		}
-
-		// Check for strconv.Xxx(c.Query/Param(...))
-		a.inferTypeFromStrconv(call, info)
-		return true
-	})
-}
-
-// inferTypeFromStrconv checks if a call is strconv.Atoi/ParseInt/ParseBool/ParseFloat
-// wrapping a c.Query/c.Param/c.DefaultQuery call, and updates the parameter's GoType.
-func (a *HandlerAnalyzer) inferTypeFromStrconv(call *ast.CallExpr, info *HandlerInfo) {
-	sel, ok := call.Fun.(*ast.SelectorExpr)
-	if !ok {
-		return
-	}
-	x, ok := sel.X.(*ast.Ident)
-	if !ok || x.Name != "strconv" {
-		return
-	}
-
-	// Determine target type from function name
-	targetType := ""
-	switch sel.Sel.Name {
-	case "Atoi", "ParseInt":
-		targetType = goTypeInteger
-	case "ParseBool":
-		targetType = goTypeBoolean
-	case "ParseFloat":
-		targetType = goTypeNumber
-	default:
-		return
-	}
-
-	// Find c.Query/c.Param/c.DefaultQuery in the arguments
-	if len(call.Args) < 1 {
-		return
-	}
-
-	paramName := a.findContextQueryParam(call.Args[0])
-	if paramName != "" {
-		a.updateParamType(info, paramName, targetType)
-	}
-}
-
-// inferBoolFromComparison checks for patterns like c.Query("x") == "true"
-// or verbose == "true" (where verbose was assigned from c.Query)
-// which suggest the parameter is boolean-typed.
-func (a *HandlerAnalyzer) inferBoolFromComparison(n ast.Node, info *HandlerInfo) {
-	binExpr, ok := n.(*ast.BinaryExpr)
-	if !ok || binExpr.Op.String() != "==" {
-		return
-	}
-
-	// Check if one side is "true"/"false" literal
-	var exprSide ast.Expr
-	var literalSide ast.Expr
-
-	litX, okX := binExpr.X.(*ast.BasicLit)
-	litY, okY := binExpr.Y.(*ast.BasicLit)
-	if okX && litX.Kind == token.STRING {
-		literalSide = binExpr.X
-		exprSide = binExpr.Y
-	} else if okY && litY.Kind == token.STRING {
-		literalSide = binExpr.Y
-		exprSide = binExpr.X
-	} else {
-		return
-	}
-
-	lit, ok := literalSide.(*ast.BasicLit)
-	if !ok || lit.Kind != token.STRING {
-		return
-	}
-	value := strings.Trim(lit.Value, `"`)
-	if value != "true" && value != "false" {
-		return
-	}
-
-	// Check if the expression side is a c.Query/c.Param call
-	paramName := a.findContextQueryParam(exprSide)
-	if paramName != "" {
-		a.updateParamType(info, paramName, goTypeBoolean)
-		return
-	}
-
-	// Also check if the expression side is a variable assigned from c.Query/c.Param
-	// We check the existing query params — if a variable name matches a query param name,
-	// this comparison suggests the param is boolean
-	if ident, ok := exprSide.(*ast.Ident); ok {
-		for i := range info.QueryParams {
-			if info.QueryParams[i].Name == ident.Name {
-				info.QueryParams[i].GoType = goTypeBoolean
-				slog.Debug("Inferred parameter type from comparison", "param", ident.Name, "type", goTypeBoolean)
-				return
-			}
-		}
-	}
-}
-
-// findContextQueryParam checks if an expression is c.Query("name") or c.Param("name")
-// and returns the parameter name. Recursively unwraps simple wrappers (e.g., type casts).
-func (a *HandlerAnalyzer) findContextQueryParam(expr ast.Expr) string {
-	call, ok := expr.(*ast.CallExpr)
-	if !ok {
-		return ""
-	}
-	sel, ok := call.Fun.(*ast.SelectorExpr)
-	if !ok {
-		return ""
-	}
-	if _, ok := sel.X.(*ast.Ident); !ok {
-		return ""
-	}
-
-	switch sel.Sel.Name {
-	case "Query", "DefaultQuery", "Param":
-		if len(call.Args) >= 1 {
-			return extractStringLiteral(call.Args[0])
-		}
-	}
-	return ""
-}
-
-// updateParamType finds an existing parameter by name and updates its GoType.
-// Searches query, path, and header params. Does nothing if not found.
-func (a *HandlerAnalyzer) updateParamType(info *HandlerInfo, paramName, targetType string) {
-	for i := range info.QueryParams {
-		if info.QueryParams[i].Name == paramName {
-			info.QueryParams[i].GoType = targetType
-			slog.Debug("Inferred parameter type", "param", paramName, "type", targetType)
-			return
-		}
-	}
-	for i := range info.PathParams {
-		if info.PathParams[i].Name == paramName {
-			info.PathParams[i].GoType = targetType
-			slog.Debug("Inferred parameter type", "param", paramName, "type", targetType)
-			return
-		}
-	}
-	for i := range info.HeaderParams {
-		if info.HeaderParams[i].Name == paramName {
-			info.HeaderParams[i].GoType = targetType
-			slog.Debug("Inferred parameter type", "param", paramName, "type", targetType)
-			return
-		}
 	}
 }
