+++ internal/extractor/gin/helper_tracer_test.go
@@ -76,125 +76,6 @@ func GetProject(c *gin.Context) {
 	}
 }
 
-func TestParameterTypeInference_Strconv(t *testing.T) {
-	src := `package main
-
-import (
-	"github.com/gin-gonic/gin"
-	"strconv"
-)
-
-func listItems(c *gin.Context) {
-	offset, _ := strconv.Atoi(c.Query("offset"))
-	limit := c.Query("limit")
-	_ = offset
-	_ = limit
-	c.JSON(200, gin.H{"ok": true})
-}
-`
-	fset := token.NewFileSet()
-	file, err := parser.ParseFile(fset, "test.go", src, parser.ParseComments)
-	if err != nil {
-		t.Fatalf("failed to parse: %v", err)
-	}
-	files := map[string]*ast.File{"test.go": file}
-
-	analyzer := NewHandlerAnalyzer(fset, files)
-
-	var fn *ast.FuncDecl
-	for _, decl := range file.Decls {
-		if f, ok := decl.(*ast.FuncDecl); ok && f.Name.Name == "listItems" {
-			fn = f
-			break
-		}
-	}
-	if fn == nil {
-		t.Fatal("listItems handler not found")
-	}
-
-	info, err := analyzer.AnalyzeHandler(fn)
-	if err != nil {
-		t.Fatalf("failed to analyze: %v", err)
-	}
-
-	// offset should be inferred as integer from strconv.Atoi
-	foundOffset := false
-	for _, p := range info.QueryParams {
-		if p.Name == "offset" {
-			foundOffset = true
-			if p.GoType != "integer" {
-				t.Errorf("expected offset GoType 'integer', got %q", p.GoType)
-			}
-		}
-	}
-	if !foundOffset {
-		t.Error("expected 'offset' query param")
-	}
-
-	// limit should remain string (no strconv wrapping)
-	foundLimit := false
-	for _, p := range info.QueryParams {
-		if p.Name == "limit" {
-			foundLimit = true
-			if p.GoType != "string" {
-				t.Errorf("expected limit GoType 'string', got %q", p.GoType)
-			}
-		}
-	}
-	if !foundLimit {
-		t.Error("expected 'limit' query param")
-	}
-}
-
-func TestParameterTypeInference_BoolComparison(t *testing.T) {
-	src := `package main
-
-import "github.com/gin-gonic/gin"
-
-func getItems(c *gin.Context) {
-	verbose := c.Query("verbose")
-	if verbose == "true" {
-		_ = verbose
-	}
-	c.JSON(200, nil)
-}
-`
-	fset := token.NewFileSet()
-	file, err := parser.ParseFile(fset, "test.go", src, parser.ParseComments)
-	if err != nil {
-		t.Fatalf("failed to parse: %v", err)
-	}
-	files := map[string]*ast.File{"test.go": file}
-
-	analyzer := NewHandlerAnalyzer(fset, files)
-
-	var fn *ast.FuncDecl
-	for _, decl := range file.Decls {
-		if f, ok := decl.(*ast.FuncDecl); ok && f.Name.Name == "getItems" {
-			fn = f
-			break
-		}
-	}
-
-	info, err := analyzer.AnalyzeHandler(fn)
-	if err != nil {
-		t.Fatalf("failed to analyze: %v", err)
-	}
-
-	foundVerbose := false
-	for _, p := range info.QueryParams {
-		if p.Name == "verbose" {
-			foundVerbose = true
-			if p.GoType != "boolean" {
-				t.Errorf("expected verbose GoType 'boolean', got %q", p.GoType)
-			}
-		}
-	}
-	if !foundVerbose {
-		t.Error("expected 'verbose' query param")
-	}
-}
-
 func TestIsKnownResponseHelper(t *testing.T) {
 	tests := []struct {
 		name     string
@@ -382,28 +263,5 @@ func Handler(c *gin.Context) {
 	// Should only have c.JSON response
 	if len(info.Responses) != 1 {
 		t.Errorf("expected 1 response (c.JSON only), got %d: %+v", len(info.Responses), info.Responses)
-	}
-}
-
-func TestIsErrorType(t *testing.T) {
-	if isErrorType(&ast.Ident{Name: "nil"}, nil) {
-		t.Error("expected nil ident to NOT be error type")
-	}
-	if isErrorType(&ast.Ident{Name: "e"}, nil) {
-		t.Error("expected single-letter 'e' to NOT be error type")
-	}
-	if !isErrorType(&ast.Ident{Name: "err"}, nil) {
-		t.Error("expected 'err' ident to be error type")
-	}
-	if isErrorType(&ast.Ident{Name: "data"}, nil) {
-		t.Error("expected 'data' ident to NOT be error type")
-	}
-	varTypeMap := map[string]string{"myErr": "error"}
-	if !isErrorType(&ast.Ident{Name: "myErr"}, varTypeMap) {
-		t.Error("expected variable with error type to be error type")
-	}
-	varTypeMap2 := map[string]string{"data": "string"}
-	if isErrorType(&ast.Ident{Name: "data"}, varTypeMap2) {
-		t.Error("expected variable with string type to NOT be error type")
 	}
 }
